Finished writing heuristic-generation function.
However, error checking reveals that cube manipulation functions are not doing exactly what they should be doing...

It appears some manipulations are causing invalid cubie position indications (eg, bits 5-7 of the representation are non-zero.)

Not sure how this is happening. Possibly might be a defect in the given
document for representation of the cube.